Tentative patch for broadcom-wl 6.30.223.271 driver for Linux 6.15-rc1
Due to the fact that the driver is unmaintained, and can't be adapted
to the current Linux kernel security standards (e.g. IBT), as it has a
proprietary blob; and this is likely to be an ever-growing security
risk, I have added a warning when the module is loaded.
Use at your own risk; no warranty of any kind is provided.
See also the objtool changes on the Makefile below.

The changes to replace EXTRA_CFLAGS and EXTRA_LDFLAGS with ccflags-y
and ldflags-y are rel. commit "kbuild: remove EXTRA_*FLAGS support"
(Masahiro Yamada, 6 Feb 2025), according to which they have been
deprecated since 2007, so no need to add any fallback for old kernels.

NB: If the package build also contains references to EXTRA_*FLAGS
(e.g. Arch Linux's broadcom-wl-dkms PKGBUILD), replace those as well!